LIBS: updated glm

+++ contrib/libs/glm/glm/gtc/packing.inl
@@ -650,7 +650,7 @@ namespace detail
 		detail::u9u9u9e5 Unpack;
 		Unpack.pack = v;
 
-		return vec3(Unpack.data.x, Unpack.data.y, Unpack.data.z) * pow(2.0f, Unpack.data.w - 15.f - 9.f);
+		return vec3(Unpack.data.x, Unpack.data.y, Unpack.data.z) * pow(2.0f, static_cast<float>(Unpack.data.w) - 15.f - 9.f);
 	}
 
 	// Based on Brian Karis http://graphicrants.blogspot.fr/2009/04/rgbm-color-encoding.html
